first step to define agent interactions
